+++ examples/1.1-check-solarium-and-ping.php
@@ -19,7 +19,7 @@ try {
     echo '<br/><pre>';
     var_dump($result->getData());
     echo '</pre>';
-} catch (Solarium\Exception $e) {
+} catch (Exception $e) {
     echo 'Ping query failed';
 }