+++ src/Exception/InvalidArgumentException.php
@@ -5,6 +5,6 @@ namespace Solarium\Exception;
 /**
  * InvalidArgument exception for Solarium classes.
  */
-class InvalidArgumentException extends \InvalidArgumentException implements ExceptionInterface
+class InvalidArgumentException extends \InvalidArgumentException implements LogicExceptionInterface
 {
 }