+++ src/Exception/UnexpectedValueException.php
@@ -5,6 +5,6 @@ namespace Solarium\Exception;
 /**
  * UnexpectedValueException exception for Solarium classes.
  */
-class UnexpectedValueException extends \UnexpectedValueException implements ExceptionInterface
+class UnexpectedValueException extends \UnexpectedValueException implements RuntimeExceptionInterface
 {
 }