+++ tests/QueryType/MoreLikeThis/QueryTest.php
@@ -5,6 +5,8 @@ namespace Solarium\Tests\QueryType\MoreLikeThis;
 use PHPUnit\Framework\TestCase;
 use Solarium\Component\MoreLikeThis;
 use Solarium\Core\Client\Client;
+use Solarium\Exception\InvalidArgumentException;
+use Solarium\Exception\OutOfBoundsException;
 use Solarium\QueryType\MoreLikeThis\Query;
 use Solarium\QueryType\Select\Query\FilterQuery;
 
@@ -249,7 +251,7 @@ class QueryTest extends TestCase
         $fq = new FilterQuery();
         $fq->setQuery('category:1');
 
-        $this->expectException('Solarium\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->query->addFilterQuery($fq);
     }
 
@@ -262,7 +264,7 @@ class QueryTest extends TestCase
         $fq2->setKey('fq1')->setQuery('category:2');
 
         $this->query->addFilterQuery($fq1);
-        $this->expectException('Solarium\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->query->addFilterQuery($fq2);
     }
 
@@ -476,7 +478,7 @@ class QueryTest extends TestCase
 
     public function testGetInvalidComponentAutoload()
     {
-        $this->expectException('Solarium\Exception\OutOfBoundsException');
+        $this->expectException(OutOfBoundsException::class);
         $this->query->getComponent('invalid', true);
     }